ui/properties: extract Results View, Sidebar and Places tabs
