+++ panel-plugin/meson.build
@@ -40,6 +40,9 @@ plugin_sources = [
   'ui/properties/search-bar.cpp',
   'ui/properties/search-bar-aliases.cpp',
   'ui/properties/search-bar-actions.cpp',
+  'ui/properties/results-view.cpp',
+  'ui/properties/sidebar.cpp',
+  'ui/properties/places.cpp',
   'slot.h',
   'window.cpp',
   xfce_revision_h,
