+++ panel-plugin/settings-dialog.cpp
@@ -732,703 +732,8 @@ void SettingsDialog::apply_unified_bar_sensitivity()
 }
 
 //-----------------------------------------------------------------------------
-// Per-tab builders for the Properties dialog. Each init_*_tab() returns a
-// fully wired scrolled container ready to be added to the dialog's stack.
-// Builders for the General, User/Session and Search Bar tabs live in their
-// own translation units under ui/properties/.
+// Per-tab builders for the Properties dialog live in their own translation
+// units under ui/properties/. Each init_*_tab() returns a fully wired
+// scrolled container ready to be added to the dialog's stack.
 //-----------------------------------------------------------------------------
 
-/* init_app_grid_tab:
- *
- * Builds the App Grid tab in the 003-properties-refactor 5-tab dictionary.
- * Sections (top-to-bottom, FR-040):
- *   1. View              — Show applications as (Icons / List / Tree) icon-radios.
- *   2. Layout            — Grid density, application icon size, show flags.
- *   3. Opacity           — App box opacity (enable-when-docked).
- *
- * Sub-enable rules per view-mode (FR-041 / FR-042): grid-density and
- * launcher-icon-size are sensitive only when view-mode == icons;
- * launcher-show-description only when view-mode == list.
- *
- * Returns: a scrolled container ready to be packed into the dialog's stack.
- */
-GtkWidget* SettingsDialog::init_app_grid_tab()
-{
-	GtkBox* page = GTK_BOX(gtk_box_new(GTK_ORIENTATION_VERTICAL, 18));
-	gtk_container_set_border_width(GTK_CONTAINER(page), 12);
-
-	GtkSizeGroup* label_size_group = gtk_size_group_new(GTK_SIZE_GROUP_HORIZONTAL);
-	GtkSizeGroup* control_size_group = gtk_size_group_new(GTK_SIZE_GROUP_HORIZONTAL);
-
-	// =========================================================================
-	// 1. View section — three exclusive icon-buttons (FR-040).
-	// =========================================================================
-	GtkGrid* view_table = GTK_GRID(gtk_grid_new());
-	gtk_grid_set_column_spacing(view_table, 12);
-	gtk_grid_set_row_spacing(view_table, 6);
-
-	GtkWidget* view_frame = make_aligned_frame(_("View"), GTK_WIDGET(view_table));
-	gtk_box_pack_start(page, view_frame, false, false, 0);
-
-	GtkButtonBox* view_box = GTK_BUTTON_BOX(gtk_button_box_new(GTK_ORIENTATION_HORIZONTAL));
-	gtk_widget_set_halign(GTK_WIDGET(view_box), GTK_ALIGN_CENTER);
-	gtk_button_box_set_layout(view_box, GTK_BUTTONBOX_EXPAND);
-	gtk_grid_attach(view_table, GTK_WIDGET(view_box), 0, 0, 2, 1);
-	gtk_widget_set_margin_bottom(GTK_WIDGET(view_box), 6);
-
-	m_show_as_icons = gtk_radio_button_new_with_mnemonic(nullptr, _("Show as _icons"));
-	{
-		const gchar* icons[] = { "view-list-icons", "view-grid", nullptr };
-		GIcon* gicon = g_themed_icon_new_from_names(const_cast<gchar**>(icons), -1);
-		gtk_button_set_image(GTK_BUTTON(m_show_as_icons), gtk_image_new_from_gicon(gicon, GTK_ICON_SIZE_DND));
-		g_object_unref(gicon);
-	}
-	gtk_button_set_image_position(GTK_BUTTON(m_show_as_icons), GTK_POS_TOP);
-	gtk_button_set_always_show_image(GTK_BUTTON(m_show_as_icons), true);
-	gtk_toggle_button_set_mode(GTK_TOGGLE_BUTTON(m_show_as_icons), false);
-	gtk_box_pack_start(GTK_BOX(view_box), m_show_as_icons, true, true, 0);
-
-	m_show_as_list = gtk_radio_button_new_with_mnemonic_from_widget(
-		GTK_RADIO_BUTTON(m_show_as_icons), _("Show as lis_t"));
-	{
-		const gchar* icons[] = { "view-list-compact", "view-list-details", "view-list", nullptr };
-		GIcon* gicon = g_themed_icon_new_from_names(const_cast<gchar**>(icons), -1);
-		gtk_button_set_image(GTK_BUTTON(m_show_as_list), gtk_image_new_from_gicon(gicon, GTK_ICON_SIZE_DND));
-		g_object_unref(gicon);
-	}
-	gtk_button_set_image_position(GTK_BUTTON(m_show_as_list), GTK_POS_TOP);
-	gtk_button_set_always_show_image(GTK_BUTTON(m_show_as_list), true);
-	gtk_toggle_button_set_mode(GTK_TOGGLE_BUTTON(m_show_as_list), false);
-	gtk_box_pack_start(GTK_BOX(view_box), m_show_as_list, true, true, 0);
-
-	m_show_as_tree = gtk_radio_button_new_with_mnemonic_from_widget(
-		GTK_RADIO_BUTTON(m_show_as_list), _("Show as t_ree"));
-	{
-		const gchar* icons[] = { "view-list-tree", "view-list-details", "pan-end", nullptr };
-		GIcon* gicon = g_themed_icon_new_from_names(const_cast<gchar**>(icons), -1);
-		gtk_button_set_image(GTK_BUTTON(m_show_as_tree), gtk_image_new_from_gicon(gicon, GTK_ICON_SIZE_DND));
-		g_object_unref(gicon);
-	}
-	gtk_button_set_image_position(GTK_BUTTON(m_show_as_tree), GTK_POS_TOP);
-	gtk_button_set_always_show_image(GTK_BUTTON(m_show_as_tree), true);
-	gtk_toggle_button_set_mode(GTK_TOGGLE_BUTTON(m_show_as_tree), false);
-	gtk_box_pack_start(GTK_BOX(view_box), m_show_as_tree, true, true, 0);
-
-	if (m_settings->view_mode == Settings::ViewAsIcons)
-		gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(m_show_as_icons), true);
-	else if (m_settings->view_mode == Settings::ViewAsTree)
-		gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(m_show_as_tree), true);
-	else
-		gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(m_show_as_list), true);
-
-	// =========================================================================
-	// 2. Layout section — grid density, icon size, show flags.
-	// =========================================================================
-	GtkGrid* layout_table = GTK_GRID(gtk_grid_new());
-	gtk_grid_set_column_spacing(layout_table, 12);
-	gtk_grid_set_row_spacing(layout_table, 6);
-
-	GtkWidget* layout_frame = make_aligned_frame(_("Layout"), GTK_WIDGET(layout_table));
-	gtk_box_pack_start(page, layout_frame, false, false, 0);
-
-	int layout_row = 0;
-
-	// Grid density (icons-only sub-enable)
-	GtkWidget* density_label = gtk_label_new_with_mnemonic(_("Grid _density:"));
-	gtk_widget_set_halign(density_label, GTK_ALIGN_START);
-	gtk_grid_attach(layout_table, density_label, 0, layout_row, 1, 1);
-
-	m_grid_density_combo = gtk_combo_box_text_new();
-	gtk_combo_box_text_append(GTK_COMBO_BOX_TEXT(m_grid_density_combo), "low", _("Low"));
-	gtk_combo_box_text_append(GTK_COMBO_BOX_TEXT(m_grid_density_combo), "medium", _("Medium"));
-	gtk_combo_box_text_append(GTK_COMBO_BOX_TEXT(m_grid_density_combo), "high", _("High"));
-	gtk_combo_box_set_active_id(GTK_COMBO_BOX(m_grid_density_combo),
-		static_cast<const gchar*>(m_settings->grid_density));
-	gtk_widget_set_halign(m_grid_density_combo, GTK_ALIGN_START);
-	gtk_grid_attach(layout_table, m_grid_density_combo, 1, layout_row, 1, 1);
-	gtk_label_set_mnemonic_widget(GTK_LABEL(density_label), m_grid_density_combo);
-	gtk_size_group_add_widget(label_size_group, density_label);
-	gtk_size_group_add_widget(control_size_group, m_grid_density_combo);
-	++layout_row;
-
-	connect(m_grid_density_combo, "changed",
-		[this](GtkComboBox* combo)
-		{
-			const gchar* val = gtk_combo_box_get_active_id(combo);
-			if (!val)
-				return;
-			m_settings->grid_density = val;
-			m_plugin->reload_menu();
-			refresh_customized_indicator();
-		});
-
-	// Application icon size (icons-only sub-enable)
-	GtkWidget* icon_size_label = gtk_label_new_with_mnemonic(_("Application icon si_ze:"));
-	gtk_widget_set_halign(icon_size_label, GTK_ALIGN_START);
-	gtk_grid_attach(layout_table, icon_size_label, 0, layout_row, 1, 1);
-
-	m_item_icon_size = gtk_combo_box_text_new();
-	gtk_widget_set_halign(m_item_icon_size, GTK_ALIGN_START);
-	const auto icon_sizes = IconSize::get_strings();
-	for (const auto& icon_size : icon_sizes)
-		gtk_combo_box_text_append_text(GTK_COMBO_BOX_TEXT(m_item_icon_size), icon_size.c_str());
-	gtk_combo_box_set_active(GTK_COMBO_BOX(m_item_icon_size), m_settings->launcher_icon_size + 1);
-	gtk_grid_attach(layout_table, m_item_icon_size, 1, layout_row, 1, 1);
-	gtk_label_set_mnemonic_widget(GTK_LABEL(icon_size_label), m_item_icon_size);
-	gtk_size_group_add_widget(label_size_group, icon_size_label);
-	gtk_size_group_add_widget(control_size_group, m_item_icon_size);
-	++layout_row;
-
-	connect(m_item_icon_size, "changed",
-		[this](GtkComboBox* combo)
-		{
-			m_settings->launcher_icon_size = gtk_combo_box_get_active(combo) - 1;
-		});
-
-	// NOTE: launcher_show_name stores "show the real (non-generic) name"; the
-	// checkbox is therefore inverted — checked means "Show generic" (false).
-	m_show_generic_names = gtk_check_button_new_with_mnemonic(_("Show generic application _names"));
-	gtk_grid_attach(layout_table, m_show_generic_names, 0, layout_row, 2, 1);
-	gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(m_show_generic_names), !m_settings->launcher_show_name);
-	++layout_row;
-
-	connect(m_show_generic_names, "toggled",
-		[this](GtkToggleButton* button)
-		{
-			m_settings->launcher_show_name = !gtk_toggle_button_get_active(button);
-			m_plugin->reload_menu();
-		});
-
-	m_show_tooltips = gtk_check_button_new_with_mnemonic(_("Show application too_ltips"));
-	gtk_grid_attach(layout_table, m_show_tooltips, 0, layout_row, 2, 1);
-	gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(m_show_tooltips), m_settings->launcher_show_tooltip);
-	++layout_row;
-
-	connect(m_show_tooltips, "toggled",
-		[this](GtkToggleButton* button)
-		{
-			m_settings->launcher_show_tooltip = gtk_toggle_button_get_active(button);
-		});
-
-	// Show descriptions — list-only sub-enable.
-	m_show_descriptions = gtk_check_button_new_with_mnemonic(_("Show application _descriptions"));
-	gtk_grid_attach(layout_table, m_show_descriptions, 0, layout_row, 2, 1);
-	gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(m_show_descriptions), m_settings->launcher_show_description);
-	++layout_row;
-
-	connect(m_show_descriptions, "toggled",
-		[this](GtkToggleButton* button)
-		{
-			m_settings->launcher_show_description = gtk_toggle_button_get_active(button);
-			m_plugin->reload_menu();
-		});
-
-	// Apply the view-mode sub-enables now and on every toggle.
-	auto apply_view_mode_sub_enables = [this, density_label, icon_size_label]()
-	{
-		const bool is_icons = (m_settings->view_mode == Settings::ViewAsIcons);
-		const bool is_list  = (m_settings->view_mode == Settings::ViewAsList);
-		gtk_widget_set_sensitive(m_grid_density_combo, is_icons);
-		gtk_widget_set_sensitive(density_label, is_icons);
-		gtk_widget_set_sensitive(m_item_icon_size, is_icons);
-		gtk_widget_set_sensitive(icon_size_label, is_icons);
-		gtk_widget_set_sensitive(m_show_descriptions, is_list);
-	};
-	apply_view_mode_sub_enables();
-
-	connect(m_show_as_icons, "toggled",
-		[this, apply_view_mode_sub_enables](GtkToggleButton* button)
-		{
-			if (!gtk_toggle_button_get_active(button))
-				return;
-			m_settings->view_mode = Settings::ViewAsIcons;
-			apply_view_mode_sub_enables();
-			m_plugin->reload_menu();
-			refresh_customized_indicator();
-		});
-
-	connect(m_show_as_list, "toggled",
-		[this, apply_view_mode_sub_enables](GtkToggleButton* button)
-		{
-			if (!gtk_toggle_button_get_active(button))
-				return;
-			m_settings->view_mode = Settings::ViewAsList;
-			apply_view_mode_sub_enables();
-			m_plugin->reload_menu();
-			refresh_customized_indicator();
-		});
-
-	connect(m_show_as_tree, "toggled",
-		[this, apply_view_mode_sub_enables](GtkToggleButton* button)
-		{
-			if (!gtk_toggle_button_get_active(button))
-				return;
-			m_settings->view_mode = Settings::ViewAsTree;
-			apply_view_mode_sub_enables();
-			m_plugin->reload_menu();
-			refresh_customized_indicator();
-		});
-
-	// =========================================================================
-	// 3. Opacity section — App box opacity (FR-044, enable-when-docked).
-	// =========================================================================
-	GtkGrid* opacity_table = GTK_GRID(gtk_grid_new());
-	gtk_grid_set_column_spacing(opacity_table, 12);
-	gtk_grid_set_row_spacing(opacity_table, 6);
-
-	GtkWidget* opacity_frame = make_aligned_frame(_("Opacity"), GTK_WIDGET(opacity_table));
-	gtk_box_pack_start(page, opacity_frame, false, false, 0);
-
-	GtkWidget* apps_op_label = gtk_label_new_with_mnemonic(_("App _box opacity:"));
-	gtk_widget_set_halign(apps_op_label, GTK_ALIGN_START);
-	gtk_grid_attach(opacity_table, apps_op_label, 0, 0, 1, 1);
-
-	m_apps_opacity = gtk_scale_new_with_range(GTK_ORIENTATION_HORIZONTAL, 0.0, 100.0, 1.0);
-	gtk_widget_set_hexpand(m_apps_opacity, true);
-	gtk_scale_set_value_pos(GTK_SCALE(m_apps_opacity), GTK_POS_RIGHT);
-	gtk_range_set_value(GTK_RANGE(m_apps_opacity), m_settings->apps_opacity);
-	gtk_grid_attach(opacity_table, m_apps_opacity, 1, 0, 1, 1);
-	gtk_label_set_mnemonic_widget(GTK_LABEL(apps_op_label), m_apps_opacity);
-	gtk_size_group_add_widget(label_size_group, apps_op_label);
-
-	connect(m_apps_opacity, "value-changed",
-		[this](GtkRange* range)
-		{
-			m_settings->apps_opacity = static_cast<int>(gtk_range_get_value(range));
-			m_plugin->reload_menu();
-			refresh_customized_indicator();
-		});
-
-	// Layout-mode-sensitive: enable only in docked.
-	m_layout_enable_when_docked.push_back(m_apps_opacity);
-	m_layout_enable_when_docked.push_back(apps_op_label);
-
-	return wrap_in_scrolled(GTK_WIDGET(page));
-}
-
-/* init_sidebar_tab:
- *
- * Builds the Sidebar tab in the 003-properties-refactor 5-tab dictionary.
- * Sections (top-to-bottom, FR-050):
- *   1. Visuals      — show category names, icon size, sidebar opacity.
- *   2. Position     — sidebar position (left/right/hidden).
- *   3. Behaviour    — hover-switch, sort categories, default category.
- *   4. Recently used — max items, include favorites.
- *
- * Sub-enable rule for category-show-name: greyed when sidebar-position
- * is not one of {left, right} (data-model E-6 §SidebarLeftRight).
- *
- * Returns: a scrolled container ready to be packed into the dialog's stack.
- */
-GtkWidget* SettingsDialog::init_sidebar_tab()
-{
-	GtkBox* page = GTK_BOX(gtk_box_new(GTK_ORIENTATION_VERTICAL, 18));
-	gtk_container_set_border_width(GTK_CONTAINER(page), 12);
-
-	GtkSizeGroup* label_size_group = gtk_size_group_new(GTK_SIZE_GROUP_HORIZONTAL);
-	GtkSizeGroup* control_size_group = gtk_size_group_new(GTK_SIZE_GROUP_HORIZONTAL);
-
-	// =========================================================================
-	// 1. Visuals section
-	// =========================================================================
-	GtkGrid* visuals_table = GTK_GRID(gtk_grid_new());
-	gtk_grid_set_column_spacing(visuals_table, 12);
-	gtk_grid_set_row_spacing(visuals_table, 6);
-
-	GtkWidget* visuals_frame = make_aligned_frame(_("Visuals"), GTK_WIDGET(visuals_table));
-	gtk_box_pack_start(page, visuals_frame, false, false, 0);
-
-	int v_row = 0;
-
-	m_show_category_names = gtk_check_button_new_with_mnemonic(_("Show cate_gory names"));
-	gtk_grid_attach(visuals_table, m_show_category_names, 0, v_row, 2, 1);
-	gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(m_show_category_names),
-		m_settings->category_show_name);
-	++v_row;
-
-	connect(m_show_category_names, "toggled",
-		[this](GtkToggleButton* button)
-		{
-			m_settings->category_show_name = gtk_toggle_button_get_active(button);
-		});
-
-	// Category icon size
-	GtkWidget* cat_size_label = gtk_label_new_with_mnemonic(_("Categ_ory icon size:"));
-	gtk_widget_set_halign(cat_size_label, GTK_ALIGN_START);
-	gtk_grid_attach(visuals_table, cat_size_label, 0, v_row, 1, 1);
-
-	m_category_icon_size = gtk_combo_box_text_new();
-	gtk_widget_set_halign(m_category_icon_size, GTK_ALIGN_START);
-	const auto cat_icon_sizes = IconSize::get_strings();
-	for (const auto& s : cat_icon_sizes)
-		gtk_combo_box_text_append_text(GTK_COMBO_BOX_TEXT(m_category_icon_size), s.c_str());
-	gtk_combo_box_set_active(GTK_COMBO_BOX(m_category_icon_size),
-		m_settings->category_icon_size + 1);
-	gtk_grid_attach(visuals_table, m_category_icon_size, 1, v_row, 1, 1);
-	gtk_label_set_mnemonic_widget(GTK_LABEL(cat_size_label), m_category_icon_size);
-	gtk_size_group_add_widget(label_size_group, cat_size_label);
-	gtk_size_group_add_widget(control_size_group, m_category_icon_size);
-	++v_row;
-
-	connect(m_category_icon_size, "changed",
-		[this](GtkComboBox* combo)
-		{
-			m_settings->category_icon_size = gtk_combo_box_get_active(combo) - 1;
-		});
-
-	// Sidebar opacity (renamed from "Category opacity"; enable-when-docked).
-	GtkWidget* side_op_label = gtk_label_new_with_mnemonic(_("_Sidebar opacity:"));
-	gtk_widget_set_halign(side_op_label, GTK_ALIGN_START);
-	gtk_grid_attach(visuals_table, side_op_label, 0, v_row, 1, 1);
-
-	m_categories_opacity = gtk_scale_new_with_range(GTK_ORIENTATION_HORIZONTAL, 0.0, 100.0, 1.0);
-	gtk_widget_set_hexpand(m_categories_opacity, true);
-	gtk_scale_set_value_pos(GTK_SCALE(m_categories_opacity), GTK_POS_RIGHT);
-	gtk_range_set_value(GTK_RANGE(m_categories_opacity), m_settings->categories_opacity);
-	gtk_grid_attach(visuals_table, m_categories_opacity, 1, v_row, 1, 1);
-	gtk_label_set_mnemonic_widget(GTK_LABEL(side_op_label), m_categories_opacity);
-	gtk_size_group_add_widget(label_size_group, side_op_label);
-	++v_row;
-
-	connect(m_categories_opacity, "value-changed",
-		[this](GtkRange* range)
-		{
-			m_settings->categories_opacity = static_cast<int>(gtk_range_get_value(range));
-			m_plugin->reload_menu();
-			refresh_customized_indicator();
-		});
-
-	m_layout_enable_when_docked.push_back(m_categories_opacity);
-	m_layout_enable_when_docked.push_back(side_op_label);
-
-	// =========================================================================
-	// 2. Position section
-	// =========================================================================
-	GtkGrid* pos_table = GTK_GRID(gtk_grid_new());
-	gtk_grid_set_column_spacing(pos_table, 12);
-	gtk_grid_set_row_spacing(pos_table, 6);
-
-	GtkWidget* pos_frame = make_aligned_frame(_("Position"), GTK_WIDGET(pos_table));
-	gtk_box_pack_start(page, pos_frame, false, false, 0);
-
-	GtkWidget* side_pos_label = gtk_label_new_with_mnemonic(_("Sidebar _position:"));
-	gtk_widget_set_halign(side_pos_label, GTK_ALIGN_START);
-	gtk_grid_attach(pos_table, side_pos_label, 0, 0, 1, 1);
-
-	m_sidebar_position_combo = gtk_combo_box_text_new();
-	gtk_combo_box_text_append(GTK_COMBO_BOX_TEXT(m_sidebar_position_combo), "left", _("Left"));
-	gtk_combo_box_text_append(GTK_COMBO_BOX_TEXT(m_sidebar_position_combo), "right", _("Right"));
-	gtk_combo_box_text_append(GTK_COMBO_BOX_TEXT(m_sidebar_position_combo), "hidden", _("Hidden"));
-	gtk_combo_box_set_active_id(GTK_COMBO_BOX(m_sidebar_position_combo),
-		static_cast<const gchar*>(m_settings->sidebar_position));
-	gtk_widget_set_halign(m_sidebar_position_combo, GTK_ALIGN_START);
-	gtk_grid_attach(pos_table, m_sidebar_position_combo, 1, 0, 1, 1);
-	gtk_label_set_mnemonic_widget(GTK_LABEL(side_pos_label), m_sidebar_position_combo);
-	gtk_size_group_add_widget(label_size_group, side_pos_label);
-	gtk_size_group_add_widget(control_size_group, m_sidebar_position_combo);
-
-	auto apply_sidebar_sub_enable = [this]()
-	{
-		const gchar* p = static_cast<const gchar*>(m_settings->sidebar_position);
-		const bool lr = p && (g_strcmp0(p, "left") == 0 || g_strcmp0(p, "right") == 0);
-		gtk_widget_set_sensitive(m_show_category_names, lr);
-	};
-	apply_sidebar_sub_enable();
-
-	connect(m_sidebar_position_combo, "changed",
-		[this, apply_sidebar_sub_enable](GtkComboBox* combo)
-		{
-			const gchar* val = gtk_combo_box_get_active_id(combo);
-			if (!val)
-				return;
-			m_settings->sidebar_position = val;
-			apply_sidebar_sub_enable();
-			m_plugin->reload_menu();
-			refresh_customized_indicator();
-		});
-
-	// =========================================================================
-	// 3. Behaviour section
-	// =========================================================================
-	GtkBox* behavior_vbox = GTK_BOX(gtk_box_new(GTK_ORIENTATION_VERTICAL, 6));
-	GtkWidget* behavior_frame = make_aligned_frame(_("Behaviour"), GTK_WIDGET(behavior_vbox));
-	gtk_box_pack_start(page, behavior_frame, false, false, 0);
-
-	m_hover_switch_category = gtk_check_button_new_with_mnemonic(_("Switch categories by _hovering"));
-	gtk_box_pack_start(behavior_vbox, m_hover_switch_category, false, false, 0);
-	gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(m_hover_switch_category),
-		m_settings->category_hover_activate);
-
-	connect(m_hover_switch_category, "toggled",
-		[this](GtkToggleButton* button)
-		{
-			m_settings->category_hover_activate = gtk_toggle_button_get_active(button);
-		});
-
-	m_sort_categories = gtk_check_button_new_with_mnemonic(_("Sort ca_tegories"));
-	gtk_box_pack_start(behavior_vbox, m_sort_categories, false, false, 0);
-	gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(m_sort_categories), m_settings->sort_categories);
-
-	connect(m_sort_categories, "toggled",
-		[this](GtkToggleButton* button)
-		{
-			m_settings->sort_categories = gtk_toggle_button_get_active(button);
-			m_plugin->reload_menu();
-		});
-
-	// Default category — three radio buttons.
-	GtkBox* default_vbox = GTK_BOX(gtk_box_new(GTK_ORIENTATION_VERTICAL, 6));
-	GtkWidget* default_frame = make_aligned_frame(_("Default category"), GTK_WIDGET(default_vbox));
-	gtk_box_pack_start(page, default_frame, false, false, 0);
-
-	m_display_favorites = gtk_radio_button_new_with_mnemonic(nullptr, _("Favorites"));
-	gtk_box_pack_start(default_vbox, m_display_favorites, false, false, 0);
-
-	m_display_recent = gtk_radio_button_new_with_mnemonic_from_widget(
-		GTK_RADIO_BUTTON(m_display_favorites), _("Recently Used"));
-	gtk_box_pack_start(default_vbox, m_display_recent, false, false, 0);
-	gtk_widget_set_sensitive(m_display_recent, m_settings->recent_items_max);
-
-	m_display_applications = gtk_radio_button_new_with_mnemonic_from_widget(
-		GTK_RADIO_BUTTON(m_display_recent), _("All Applications"));
-	gtk_box_pack_start(default_vbox, m_display_applications, false, false, 0);
-
-	switch (m_settings->default_category)
-	{
-	case Settings::CategoryRecent:
-		gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(m_display_recent), true);
-		break;
-	case Settings::CategoryAll:
-		gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(m_display_applications), true);
-		break;
-	default:
-		gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(m_display_favorites), true);
-		break;
-	}
-
-	connect(m_display_favorites, "toggled",
-		[this](GtkToggleButton* button)
-		{
-			if (gtk_toggle_button_get_active(button))
-				m_settings->default_category = Settings::CategoryFavorites;
-		});
-
-	connect(m_display_recent, "toggled",
-		[this](GtkToggleButton* button)
-		{
-			if (gtk_toggle_button_get_active(button))
-				m_settings->default_category = Settings::CategoryRecent;
-		});
-
-	connect(m_display_applications, "toggled",
-		[this](GtkToggleButton* button)
-		{
-			if (gtk_toggle_button_get_active(button))
-				m_settings->default_category = Settings::CategoryAll;
-		});
-
-	// =========================================================================
-	// 4. Recently used section
-	// =========================================================================
-	GtkGrid* recent_table = GTK_GRID(gtk_grid_new());
-	gtk_grid_set_column_spacing(recent_table, 12);
-	gtk_grid_set_row_spacing(recent_table, 6);
-
-	GtkWidget* recent_frame = make_aligned_frame(_("Recently used"), GTK_WIDGET(recent_table));
-	gtk_box_pack_start(page, recent_frame, false, false, 0);
-
-	GtkWidget* max_label = gtk_label_new_with_mnemonic(_("Maximum _items:"));
-	gtk_widget_set_halign(max_label, GTK_ALIGN_START);
-	gtk_grid_attach(recent_table, max_label, 0, 0, 1, 1);
-
-	m_recent_items_max = gtk_spin_button_new_with_range(0, 100, 1);
-	gtk_spin_button_set_value(GTK_SPIN_BUTTON(m_recent_items_max), m_settings->recent_items_max);
-	gtk_grid_attach(recent_table, m_recent_items_max, 1, 0, 1, 1);
-	gtk_label_set_mnemonic_widget(GTK_LABEL(max_label), m_recent_items_max);
-	gtk_size_group_add_widget(label_size_group, max_label);
-	gtk_size_group_add_widget(control_size_group, m_recent_items_max);
-
-	connect(m_recent_items_max, "value-changed",
-		[this](GtkSpinButton* button)
-		{
-			m_settings->recent_items_max = gtk_spin_button_get_value_as_int(button);
-			const bool active = m_settings->recent_items_max;
-			gtk_widget_set_sensitive(m_display_recent, active);
-			if (!active && gtk_toggle_button_get_active(GTK_TOGGLE_BUTTON(m_display_recent)))
-				gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(m_display_favorites), true);
-		});
-
-	m_remember_favorites = gtk_check_button_new_with_mnemonic(_("Include _favorites in \"Recent\""));
-	gtk_grid_attach(recent_table, m_remember_favorites, 0, 1, 2, 1);
-	gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(m_remember_favorites),
-		m_settings->favorites_in_recent);
-
-	connect(m_remember_favorites, "toggled",
-		[this](GtkToggleButton* button)
-		{
-			m_settings->favorites_in_recent = gtk_toggle_button_get_active(button);
-		});
-
-	return wrap_in_scrolled(GTK_WIDGET(page));
-}
-
-
-
-//-----------------------------------------------------------------------------
-
-/* init_places_tab:
- *
- * Builds the Places panel (milestone 005). Seven controls bound directly to
- * the /places-prefixed Xfconf-backed Settings members. Sensitivity is gated by
- * /places/enabled at the panel level and by /places/favourites-enabled for
- * the sync dropdown (FR-037, FR-038).
- *
- * Returns: a scrolled container ready to be packed into the dialog's stack.
- */
-GtkWidget* SettingsDialog::init_places_tab()
-{
-	GtkBox* page = GTK_BOX(gtk_box_new(GTK_ORIENTATION_VERTICAL, 18));
-	gtk_container_set_border_width(GTK_CONTAINER(page), 12);
-
-	std::vector<GtkWidget*> places_dependents;
-
-	// Enable section
-	GtkGrid* enable_grid = GTK_GRID(gtk_grid_new());
-	gtk_grid_set_column_spacing(enable_grid, 12);
-	gtk_grid_set_row_spacing(enable_grid, 6);
-	gtk_box_pack_start(page, make_aligned_frame(_("Places mode"), GTK_WIDGET(enable_grid)), false, false, 0);
-
-	GtkWidget* enable_switch = gtk_switch_new();
-	GtkWidget* enable_label = gtk_label_new_with_mnemonic(_("Enable _Places"));
-	gtk_widget_set_halign(enable_label, GTK_ALIGN_START);
-	gtk_widget_set_hexpand(enable_label, true);
-	gtk_switch_set_active(GTK_SWITCH(enable_switch), m_settings->places_enabled);
-	gtk_grid_attach(enable_grid, enable_label, 0, 0, 1, 1);
-	gtk_grid_attach(enable_grid, enable_switch, 1, 0, 1, 1);
-	gtk_label_set_mnemonic_widget(GTK_LABEL(enable_label), enable_switch);
-
-	// Sections section — history/favourites toggles plus item caps.
-	GtkGrid* sections_grid = GTK_GRID(gtk_grid_new());
-	gtk_grid_set_column_spacing(sections_grid, 12);
-	gtk_grid_set_row_spacing(sections_grid, 6);
-	GtkWidget* sections_frame = make_aligned_frame(_("Sections"), GTK_WIDGET(sections_grid));
-	gtk_box_pack_start(page, sections_frame, false, false, 0);
-	places_dependents.push_back(sections_frame);
-
-	int row = 0;
-	GtkWidget* history_switch = gtk_switch_new();
-	GtkWidget* history_label = gtk_label_new_with_mnemonic(_("Enable _History section"));
-	gtk_widget_set_halign(history_label, GTK_ALIGN_START);
-	gtk_widget_set_hexpand(history_label, true);
-	gtk_switch_set_active(GTK_SWITCH(history_switch), m_settings->places_history_enabled);
-	gtk_grid_attach(sections_grid, history_label, 0, row, 1, 1);
-	gtk_grid_attach(sections_grid, history_switch, 1, row, 1, 1);
-	gtk_label_set_mnemonic_widget(GTK_LABEL(history_label), history_switch);
-	++row;
-
-	GtkWidget* fav_switch = gtk_switch_new();
-	GtkWidget* fav_label = gtk_label_new_with_mnemonic(_("Enable _Favourites section"));
-	gtk_widget_set_halign(fav_label, GTK_ALIGN_START);
-	gtk_widget_set_hexpand(fav_label, true);
-	gtk_switch_set_active(GTK_SWITCH(fav_switch), m_settings->places_favourites_enabled);
-	gtk_grid_attach(sections_grid, fav_label, 0, row, 1, 1);
-	gtk_grid_attach(sections_grid, fav_switch, 1, row, 1, 1);
-	gtk_label_set_mnemonic_widget(GTK_LABEL(fav_label), fav_switch);
-	++row;
-
-	GtkWidget* sync_label = gtk_label_new_with_mnemonic(_("Favourite _sync:"));
-	gtk_widget_set_halign(sync_label, GTK_ALIGN_START);
-	GtkWidget* sync_combo = gtk_combo_box_text_new();
-	gtk_combo_box_text_append(GTK_COMBO_BOX_TEXT(sync_combo), "meowmenu", _("MeowMenu only"));
-	gtk_combo_box_text_append(GTK_COMBO_BOX_TEXT(sync_combo), "thunar",   _("Thunar bookmarks (read-only)"));
-	gtk_combo_box_set_active_id(GTK_COMBO_BOX(sync_combo),
-			static_cast<const gchar*>(m_settings->places_favourite_sync));
-	gtk_grid_attach(sections_grid, sync_label, 0, row, 1, 1);
-	gtk_grid_attach(sections_grid, sync_combo, 1, row, 1, 1);
-	gtk_label_set_mnemonic_widget(GTK_LABEL(sync_label), sync_combo);
-	++row;
-
-	GtkWidget* max_label = gtk_label_new_with_mnemonic(_("Maximum places _items:"));
-	gtk_widget_set_halign(max_label, GTK_ALIGN_START);
-	GtkWidget* max_spin = gtk_spin_button_new_with_range(0, 30, 1);
-	gtk_spin_button_set_value(GTK_SPIN_BUTTON(max_spin), m_settings->places_max_items);
-	gtk_grid_attach(sections_grid, max_label, 0, row, 1, 1);
-	gtk_grid_attach(sections_grid, max_spin, 1, row, 1, 1);
-	gtk_label_set_mnemonic_widget(GTK_LABEL(max_label), max_spin);
-	++row;
-
-	// Behaviour section
-	GtkGrid* behaviour_grid = GTK_GRID(gtk_grid_new());
-	gtk_grid_set_column_spacing(behaviour_grid, 12);
-	gtk_grid_set_row_spacing(behaviour_grid, 6);
-	GtkWidget* behaviour_frame = make_aligned_frame(_("Behaviour"), GTK_WIDGET(behaviour_grid));
-	gtk_box_pack_start(page, behaviour_frame, false, false, 0);
-	places_dependents.push_back(behaviour_frame);
-
-	GtkWidget* remember_check = gtk_check_button_new_with_mnemonic(_("_Remember last selected mode"));
-	gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(remember_check),
-			m_settings->places_remember_last_mode);
-	gtk_grid_attach(behaviour_grid, remember_check, 0, 0, 2, 1);
-
-	GtkWidget* meta_check = gtk_check_button_new_with_mnemonic(_("Show item _metadata"));
-	gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(meta_check),
-			m_settings->places_show_metadata);
-	gtk_grid_attach(behaviour_grid, meta_check, 0, 1, 2, 1);
-
-	// Sensitivity helpers (FR-037, FR-038).
-	auto refresh_sensitivity = [=]()
-	{
-		const bool enabled = gtk_switch_get_active(GTK_SWITCH(enable_switch));
-		for (GtkWidget* w : places_dependents)
-			gtk_widget_set_sensitive(w, enabled);
-		const bool fav_enabled = enabled && gtk_switch_get_active(GTK_SWITCH(fav_switch));
-		gtk_widget_set_sensitive(sync_combo, fav_enabled);
-		gtk_widget_set_sensitive(sync_label, fav_enabled);
-	};
-	refresh_sensitivity();
-
-	// Signal wiring
-	connect(enable_switch, "state-set",
-		[this, refresh_sensitivity](GtkSwitch*, gboolean state) -> gboolean
-		{
-			m_settings->places_enabled = state;
-			refresh_sensitivity();
-			return FALSE; // let the switch update its visual state
-		});
-	connect(history_switch, "state-set",
-		[this](GtkSwitch*, gboolean state) -> gboolean
-		{
-			m_settings->places_history_enabled = state;
-			return FALSE;
-		});
-	connect(fav_switch, "state-set",
-		[this, refresh_sensitivity](GtkSwitch*, gboolean state) -> gboolean
-		{
-			m_settings->places_favourites_enabled = state;
-			refresh_sensitivity();
-			return FALSE;
-		});
-	connect(sync_combo, "changed",
-		[this](GtkComboBox* combo)
-		{
-			const gchar* val = gtk_combo_box_get_active_id(combo);
-			if (val) m_settings->places_favourite_sync = val;
-		});
-	connect(max_spin, "value-changed",
-		[this](GtkSpinButton* btn)
-		{
-			m_settings->places_max_items = gtk_spin_button_get_value_as_int(btn);
-		});
-	connect(remember_check, "toggled",
-		[this](GtkToggleButton* btn)
-		{
-			m_settings->places_remember_last_mode = gtk_toggle_button_get_active(btn);
-		});
-	connect(meta_check, "toggled",
-		[this](GtkToggleButton* btn)
-		{
-			m_settings->places_show_metadata = gtk_toggle_button_get_active(btn);
-		});
-
-	return wrap_in_scrolled(GTK_WIDGET(page));
-}
